requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.3
django-cors-headers==4.2.0
django-countries==7.5.1
django-debug-toolbar==4.1.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-smtp-ssl==1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
jsmin==3.0.1
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
